shsh

+++ part4/part4/multi_service_server.randall7.c
@@ -133,23 +133,16 @@ int main(int argc, const char * argv[]) {
             char* dummy = buf3 + 9;
             
             uint32_t again =0;
-            again = buf3[n-1];
+            again = (uint8_t) buf3[n-1];
             again <<= 8;
-            again |=  buf3[n-2];
+            again |= (uint8_t) buf3[n-2];
             again <<=8;
-            again |=  buf3[n-3];
+            again |= (uint8_t) buf3[n-3];
             again <<=8;
-            again |=  buf3[n-4];
+            again |= (uint8_t) buf3[n-4];
             
             printf("\n\nHope it works: %u\n\n\n", again);
-            
-            
-            
-            
-            
-            
-            
-            
+        
             
             uint8_t byte1 = (uint8_t)buf3[n-4];
             uint8_t byte2 = (uint8_t)buf3[n-3];
